+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added the footer row with total fees and total value to the activities table
+
 ### Changed
 
 - Upgraded _Stripe_ dependencies